fix: update path construction logic in translations
