+++ packages/api-website-builder/src/features/pages/TranslatePage/TranslatePageUseCase.ts
@@ -3,14 +3,21 @@ import { TranslatePageUseCase as UseCaseAbstraction } from "./abstractions.js";
 import { DuplicatePageRepository } from "~/features/pages/DuplicatePage/abstractions.js";
 import { GetPageByIdUseCase } from "~/features/pages/GetPageById/index.js";
 import { GetLanguageByCodeUseCase } from "@webiny/languages/exports/api/languages.js";
+import { ListLanguagesUseCase } from "@webiny/languages/exports/api/languages.js";
 import { WbPermissions } from "~/features/permissions/abstractions.js";
-import { PageNotAuthorizedError, PageTranslationError } from "~/domain/page/errors.js";
+import {
+    PageNotAuthorizedError,
+    PageTranslationError,
+    PagePersistenceError
+} from "~/domain/page/errors.js";
+import { PagePath } from "~/domain/page/PagePath.js";
 
 class TranslatePageUseCaseImpl implements UseCaseAbstraction.Interface {
     constructor(
         private permissions: WbPermissions.Interface,
         private getPageById: GetPageByIdUseCase.Interface,
         private getLanguageByCode: GetLanguageByCodeUseCase.Interface,
+        private listLanguages: ListLanguagesUseCase.Interface,
         private duplicatePageRepository: DuplicatePageRepository.Interface
     ) {}
 
@@ -20,13 +27,23 @@ class TranslatePageUseCaseImpl implements UseCaseAbstraction.Interface {
             return Result.fail(new PageNotAuthorizedError());
         }
 
-        // Validate language code
+        // Validate language code.
         const languageResult = await this.getLanguageByCode.execute(params.languageCode);
         if (languageResult.isFail()) {
             return Result.fail(new PageTranslationError(params.languageCode));
         }
 
-        // Fetch source page to resolve lineage
+        // Get all supported language codes.
+        const languagesResult = await this.listLanguages.execute();
+        if (languagesResult.isFail()) {
+            return Result.fail(
+                new PagePersistenceError(new Error("Failed to fetch supported languages"))
+            );
+        }
+
+        const supportedCodes = languagesResult.value.map(lang => lang.code);
+
+        // Fetch source page to resolve lineage.
         const getResult = await this.getPageById.execute(params.pageId);
         if (getResult.isFail()) {
             return getResult;
@@ -34,18 +51,18 @@ class TranslatePageUseCaseImpl implements UseCaseAbstraction.Interface {
 
         const sourcePage = getResult.value;
 
-        // Resolve sourcePage: always point to the root base page
+        // Resolve sourcePage: always point to the root base page.
         const resolvedSourcePageId = sourcePage.properties.sourcePage ?? sourcePage.entryId;
 
-        // Duplicate the page with translation modifications
+        // Duplicate the page with translation modifications.
         const result = await this.duplicatePageRepository.execute(
             { id: params.pageId },
             ({ duplicate }) => {
                 const originalPath: string = sourcePage.properties.path ?? "/";
-                const translatedPath =
-                    originalPath === "/"
-                        ? `/${params.languageCode}`
-                        : `/${params.languageCode}${originalPath}`;
+                const pagePath = PagePath.create(originalPath);
+                const translatedPath = pagePath
+                    .setLanguageCode(params.languageCode, supportedCodes)
+                    .toString();
 
                 duplicate.properties.language = params.languageCode;
                 duplicate.properties.sourcePage = resolvedSourcePageId;
@@ -70,6 +87,7 @@ export const TranslatePageUseCase = createImplementation({
         WbPermissions,
         GetPageByIdUseCase,
         GetLanguageByCodeUseCase,
+        ListLanguagesUseCase,
         DuplicatePageRepository
     ]
 });
